% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cumulative_table_vector.R
\name{cumulative_table_vector}
\alias{cumulative_table_vector}
\title{Cumulative Table}
\usage{
cumulative_table_vector(df, column)
}
\arguments{
\item{df, }{column.  A data frame and a column within that data frame. The column must be numeric}
}
\description{
A function to quickly create cumulative tables
}
\examples{
cumulative_table_vector(mtcars, "carb")
}
\keyword{Cumulative,}
\keyword{Table}
