% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cut_data_frame.R
\name{cut_df}
\alias{cut_df}
\title{Cut data frame}
\usage{
cut_df(df, parts)
}
\description{
A function that breaks a data frame into a smaller number of data frames of equal number of rows.
}
\examples{
x <- as_tibble(1:1000)
p <- cut_df(x,2)

}
\keyword{Break}
\keyword{data}
\keyword{frame}
