% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe_data_frame.R
\name{describe}
\alias{describe}
\title{Describe function}
\usage{
describe(data_frame)
}
\arguments{
\item{data_frame.}{A basic data frame or tibble}
}
\description{
Inspired by a function with the same name in STATA, it quickly displays information about variables
}
\examples{
describe(mtcars)
}
\keyword{describe}
