% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostic_plots.R
\name{lm_diagnostic_plots}
\alias{lm_diagnostic_plots}
\title{Linear regression diagnostic plots}
\usage{
lm_diagnostic_plots(model)
}
\description{
A function that produces ggplot equivalents of regression diagnostic plots. Works only when response is quantitative
}
\examples{
lm_diagnostic_plots(fit_1), where fit_1 = lm(y~..., data = ...)
}
\keyword{Diagnostic}
\keyword{plots}
