% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poisson_dispersion_statistic.R
\name{poisson_dispersion_statistic}
\alias{poisson_dispersion_statistic}
\title{Describe function}
\usage{
poisson_dispersion_statistic(glm_fit)
}
\arguments{
\item{glm_fit.}{An object of the type (glm,..,family = "poisson")}
}
\description{
Quickly calculate poission dispersion statistic from a GLM object
}
\examples{
poisson_dispersion_statistic(glm_fit)
}
\keyword{dispersion,}
\keyword{pearson,}
\keyword{residuals}
