% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numeric_summary.R
\name{summary_function}
\alias{summary_function}
\title{Cumulative Table}
\usage{
summary_function(vector)
}
\arguments{
\item{vector.}{A numeric vector}
}
\description{
Stata inspired summary function
}
\examples{
summary_function(mtcars$carb)
}
\keyword{summarise}
\keyword{summary,}
